+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/sessions/change/CollectionChangeDetector.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.sessions.change;
 
@@ -117,10 +115,10 @@ class CollectionChangeDetector implements ChangeDetector {
                 if (keyOne == null || keyTwo == null) {
                     return 0;
                 }
-                return keyOne.hashCode() - keyTwo.hashCode();
+                return Integer.compare(keyOne.hashCode(), keyTwo.hashCode());
             };
         } else {
-            return (o1, o2) -> o1.hashCode() - o2.hashCode();
+            return (o1, o2) -> Integer.compare(o1.hashCode(), o2.hashCode());
         }
     }
 }