+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/environment/utils/Generators.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.environment.utils;
 
@@ -38,6 +36,21 @@ public abstract class Generators {
 
     public static int randomInt(int max) {
         return RAND.nextInt(max);
+    }
+
+    /**
+     * Gets a random integer greater than one.
+     *
+     * @param max Upper bound
+     * @return Random integer
+     */
+    public static int randomPositiveInt(int max) {
+        assert max > 1;
+        int res;
+        do {
+            res = RAND.nextInt(max);
+        } while (res <= 1);
+        return res;
     }
 
     public static boolean randomBoolean() {