+++ jopa-integration-tests/src/main/java/cz/cvut/kbss/jopa/test/runner/DeleteOperationsRunner.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.test.runner;
 
@@ -462,5 +460,29 @@ public abstract class DeleteOperationsRunner extends BaseRunner {
                                   .setParameter("instance",
                                           entityQ.getUri()).getSingleResult();
         assertFalse(remains);
+    }
+
+    @Test
+    public void testRemoveDetachedWithCascadedReferenceUsingMergeAndRemove() {
+        this.em = getEntityManager("RemoveDetachedEntityWithCascadedReferenceUsingMergeAndRemove", true);
+        persist(entityH);
+        assertNotNull(em.find(OWLClassH.class, entityH.getUri()));
+        assertNotNull(em.find(OWLClassA.class, entityA.getUri()));
+        final OWLClassH h = new OWLClassH();
+        h.setUri(entityH.getUri());
+        final OWLClassA a = new OWLClassA();
+        a.setUri(entityA.getUri());
+        a.setStringAttribute(entityA.getStringAttribute());
+        a.setTypes(new HashSet<>(entityA.getTypes()));
+        h.setOwlClassA(a);
+
+        em.clear();
+        em.getTransaction().begin();
+        final OWLClassH toRemove = em.merge(h);
+        em.remove(toRemove);
+        em.getTransaction().commit();
+
+        assertNull(em.find(OWLClassH.class, entityH.getUri()));
+        assertNull(em.find(OWLClassA.class, entityA.getUri()));
     }
 }